==> 3_news_KNN.ipynb <==
## note that active directory is already set in the call to import below
## refer to line 37 and line 38 of the function file

from function_file_news_corpus import *


news_df = pd.read_pickle('final_clean_hindu_corpus.pkl')
news_df.drop(['DateTime'], axis = 1, inplace = True)

news_df.head()

## note that active directory is already set in the call to import below
## use K_m_cluster in line 624 of the function_file_news_corpus.py

K_m_df =  K_m_cluster (29, news_df, 'text', 20 )


K_m_df.head()
