==> 5_Sentiment_analysis.ipynb <==
## sentiment analysis 
## note that active directory is already set in thfunction_file_news_corpus
## Keep all files in the same active dir
## refer to line 48 and line 49 of the function file
## $ conda install -c https://conda.anaconda.org/sloria textblob
## refer to http://textblob.readthedocs.io/en/dev/install.html documentation 


from function_file_news_corpus import *
from textblob import TextBlob
def map_sentiment (text):
    analyzer = TextBlob(text)
    return analyzer.sentiment.polarity

news_df_topics = pd.read_pickle('news_df_topics_for_sentiment_analysis_final.pkl')
news_df_topics['sentiment'] = news_df_topics['text'].map(map_sentiment)

## convert data time to an amenable format, year_quarter  

def yr (text):
    return int(str(text)[0:4])

def mnth (text):
    return int(str(text)[4:6])

def qtr (text):
    text =  int(str(text)[4:6])
    if text <=3:
        return 1
    elif text <=6:
        return 2
    elif text <=9:
        return 3
    elif text <=9:
        return 3
    elif text <=12:
        return 4
    
    
def day (text):
    return int(str(text)[-2:])


news_df_topics['year'] = news_df_topics['date'].map(yr)
news_df_topics['month'] = news_df_topics['date'].map(mnth)
news_df_topics['quarter'] = news_df_topics['date'].map(qtr)
news_df_topics['day'] = news_df_topics['date'].map(day)
news_df_topics['year_quarter'] = news_df_topics['year'] + news_df_topics['quarter']/10


news_df_sentiment_topic = news_df_topics[['topic','sentiment','year_quarter','year','month','day']].copy(deep=True)

news_df_sentiment_topic['sentiment_type'] = np.where(news_df_sentiment_topic.sentiment>0, 1, -1)
news_df_sentiment_topic['sentiment_type'] = np.where(news_df_sentiment_topic.sentiment == 0, 
                                                     0,news_df_sentiment_topic.sentiment_type )
topic_key_final = pd.read_excel('topic_key_final.xlsx')
news_df_sentiment_topic = pd.merge(news_df_sentiment_topic,topic_key_final,on=['topic'])
news_df_sentiment_topic.head()

writer = pd.ExcelWriter('time_scatter_sentiment_topic.xlsx')
news_df_sentiment_topic.to_excel(writer, sheet_name='time_scatter_sentiment')
writer.save()


news_df_sentiment_topic['mean_sentiment_topic_quarter'] = news_df_sentiment_topic.groupby(['year_quarter', 'topic'])['sentiment'].transform('mean')

plotly_mean_sentiment = news_df_sentiment_topic[['topic', 'year_quarter',
                                                       'mean_sentiment_topic_quarter', 'inf_topic', 'sub_topic' ]].copy(deep=True)
         
plotly_mean_sentiment.drop_duplicates(inplace=True)

plotly_mean_sentiment = plotly_mean_sentiment.sort_values(['topic', 'year_quarter'], ascending=[1, 1])
plotly_mean_sentiment.head(n=20)


merge_year_quarter = pd.read_excel('merge_year_quarter.xlsx')

df_1 = list(plotly_mean_sentiment['topic'].unique())
df_2 =list(merge_year_quarter['year_quarter'].unique())


dfn = pd.DataFrame([(x, y) for x in df_1 for y in df_2])
dfn.rename(columns={1:'year_quarter',0 :'topic' }, inplace=True)
plotly_mean_sentiment= pd.merge(plotly_mean_sentiment,dfn,how='right',on=['year_quarter','topic'])

#plotly_mean_sentiment.fillna(0, inplace = True)
writer = pd.ExcelWriter('plotly_mean_sentiment.xlsx')
plotly_mean_sentiment.to_excel(writer, sheet_name='plotly_mean_sentiment')
writer.save()

plotly_mean_sentiment.head()

## analysis for values grouped by quarter:

news_df_sentiment_topic['mean_sentiment_topic_quarter'] = news_df_sentiment_topic.groupby(['year_quarter', 'topic'])['sentiment_type'].transform('mean')

plotly_mean_sentiment = news_df_sentiment_topic[['topic', 'year_quarter',
                                                       'mean_sentiment_topic_quarter', 'inf_topic', 'sub_topic' ]].copy(deep=True)
         
plotly_mean_sentiment.drop_duplicates(inplace=True)

plotly_mean_sentiment = plotly_mean_sentiment.sort_values(['topic', 'year_quarter'], ascending=[1, 1])
plotly_mean_sentiment.head(n=20)


merge_year_quarter = pd.read_excel('merge_year_quarter.xlsx')

df_1 = list(plotly_mean_sentiment['topic'].unique())
df_2 =list(merge_year_quarter['year_quarter'].unique())


dfn = pd.DataFrame([(x, y) for x in df_1 for y in df_2])
dfn.rename(columns={1:'year_quarter',0 :'topic' }, inplace=True)
plotly_mean_sentiment= pd.merge(plotly_mean_sentiment,dfn,how='right',on=['year_quarter','topic'])

active_dir = "/Users/anirudhsyal/Desktop/Hindu_news/mean_positive_negative"
os.chdir (active_dir )


#plotly_mean_sentiment.fillna(0, inplace = True)
writer = pd.ExcelWriter('plotly_mean_sentiment.xlsx')
plotly_mean_sentiment.to_excel(writer, sheet_name='plotly_mean_sentiment')
writer.save()

plotly_mean_sentiment.head()


## analysis for all values 

news_df = pd.read_pickle('final_clean_hindu_corpus.pkl')
news_df.drop(['DateTime'], axis = 1, inplace = True)

news_df['tokens'] = news_df['text'].map(tokenizer_tf_idf)
news_df['count_tokens'] = news_df['tokens'].map(len)
news_df_count_all = news_df[['aadhaar_count','num_date','count_tokens']].copy(deep=True)

writer = pd.ExcelWriter('news_df_count_all.xlsx')
news_df_count_all.to_excel(writer, sheet_name='news_df_count_all')
writer.save()

news_df.head()


news_df['year'] = news_df['num_date'].map(yr)
news_df['month'] = news_df['num_date'].map(mnth)
news_df['quarter'] = news_df['num_date'].map(qtr)
news_df['day'] = news_df['num_date'].map(day)
news_df['year_quarter'] = news_df['year'] + news_df_topics['quarter']/10
news_df_decriptive = news_df[['year_quarter','num_date','count_tokens','aadhaar_count' ]].copy(deep=True)
news_df_decriptive.sort_values(['year_quarter','count_tokens', 'aadhaar_count'], ascending=[1, 1, 1], inplace = True)
news_df_decriptive.head()


news_df_decriptive['aadhaar_per_100_tokens'] = (100*news_df_decriptive['aadhaar_count' ]/ news_df_decriptive['count_tokens' ]).round(3)

news_df_decriptive['mean_token_count'] = news_df_decriptive.groupby(['year_quarter'])['count_tokens'].transform('mean')

news_df_decriptive['mean_aadhaar_count'] = news_df_decriptive.groupby(['year_quarter'])['aadhaar_count'].transform('mean')

news_df_decriptive['mean_aadhaar_per_100_tokens'] = news_df_decriptive.groupby(['year_quarter'])['aadhaar_per_100_tokens'].transform('mean')

news_df_decriptive['article_occurences'] = news_df_decriptive.groupby(['year_quarter'])['aadhaar_count'].transform('count')

news_df_decriptive['aadhaar_occurences'] = news_df_decriptive.groupby(['year_quarter'])['aadhaar_count'].transform('sum')


news_df_decriptive.drop(['num_date', 'count_tokens', 'aadhaar_count', 'aadhaar_per_100_tokens'], axis = 1, inplace = True)

print(len(news_df_decriptive))
news_df_decriptive.drop_duplicates(inplace=True)
print(len(news_df_decriptive))



merge_year_quarter = pd.read_excel('merge_year_quarter.xlsx')

#df_1 = list(plotly_mean_sentiment['topic'].unique())
#df_2 =list(merge_year_quarter['year_quarter'].unique())


#dfn = pd.DataFrame([(x, y) for x in df_1 for y in df_2])
#dfn.rename(columns={1:'year_quarter',0 :'topic' }, inplace=True)
news_df_decriptive= pd.merge(news_df_decriptive,merge_year_quarter,how='right',on=['year_quarter'])
print(len(news_df_decriptive))
news_df_decriptive.sort_values(['year_quarter'], ascending=[1], inplace = True)
news_df_decriptive.fillna(0, inplace = True)
news_df_decriptive.mean_aadhaar_per_100_tokens.round(3)
news_df_decriptive.mean_token_count.round(1)
news_df_decriptive.mean_aadhaar_count.round(1)
news_df_decriptive.head()

writer = pd.ExcelWriter('news_df_decriptive_002.xlsx')
news_df_decriptive.to_excel(writer, sheet_name='news_df_decriptive')
writer.save()
